pyproject.toml
[project]
name = "helloconfig"
version = "0.2.0.dev1"
authors = [
  { name="Elchin Sarkarov", email="elchin751@gmail.com" },
]
description = "Describe config, i'll create it"
requires-python = ">=3.8"
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: POSIX",
  "Operating System :: Microsoft :: Windows",
]
dependencies = [ "dataclass_factory", "pyyaml", "libcst" ]
readme = "README.md"

[project.urls]
"Homepage" = "https://github.com/elchinchel/hello-config"
"Bug Tracker" = "https://github.com/elchinchel/hello-config/issues"

[project.optional-dependencies]
test = [ "pytest", "pytest-cov" ]

[tool.setuptools]
packages = ["helloconfig"]

[build-system]
requires = ["setuptools>=40.8.0"]
build-backend = "setuptools.build_meta"

[tool.pyright]
typeCheckingMode = "basic"
